requirements.txt
fastapi
uvicorn
python-multipart
pypdf
langchain
langchain-huggingface
langchain-community
faiss-cpu
sentence-transformers
python-dotenv
tiktoken
streamlit
requests